feat(agent): refactor agent service to integrate Sentry for error tracking and remove third-party service file
