+++ web-app/src/lib/services/agent/index.ts
@@ -1,2 +1,1 @@
 export * from "./service";
-export * from "./third-party";
